dom/broadcastchannel/tests/test_broadcastchannel_sharedWorker.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<!--
Tests of DOM BroadcastChannel in SharedWorkers
-->
<head>
  <title>Test for BroadcastChannel in SharedWorkers</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" language="javascript">

function runTests() {
  var worker = new SharedWorker("broadcastchannel_sharedWorker.js");

  var bc = new BroadcastChannel('foobar');

  worker.port.onmessage = function(event) {
    if (event.data == "READY") {
      ok(true, "SharedWorker is ready!");
      bc.postMessage('hello world from the window');
    } else {
      ok(false, "Something wrong happened");
    }
  };

  bc.onmessage = function(event) {
    is("hello world from the worker", event.data, "The message matches!");
    bc.close();
    SimpleTest.finish();
  }

  worker.port.postMessage('go');
}

SimpleTest.waitForExplicitFinish();
SpecialPowers.pushPrefEnv({"set": [["dom.broadcastChannel.enabled", true],
                                   ["dom.workers.sharedWorkers.enabled", true]]}, runTests);

</script>
</pre>
</body>
</html>
